include paths to package dirs in $registry/packages.toml
Use this info to find package paths. This makes the directory layout
of the registry self-documenting so that we don't need to rely on a
convention that might change and break everything. If need to nest
the directory structure deeper or organize by uuids instead of names
then we can easily do that without any breakage.